curator/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "curator"
version = "0.1.0"
description = "Curator service for Birds of Vinca"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Birds of Vinca Team"}
]
dependencies = [
    "flask>=3.1.0",
    "functions-framework>=3.0.0",
    "httpx>=0.28.0",
    "markupsafe>=3.0.0",
    "moviepy>=2.0.0",
    "numpy>=2.0.0",
    "openai>=2.21.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=11.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "bov-data @ file:///workspaces/birds_of_vinca/libs/bov_data"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
